factor out util
